refactor: 오류 메세지는 예외가 발생하는 클래스에서 관리하도록 변경함 - 전체적인 코드 포멧팅 수정
