+++ src/main/java/lotto/domain/lotto/LottoNumber.java
@@ -1,18 +1,17 @@
 package lotto.domain.lotto;
 
-import static lotto.view.messages.ErrorMessages.LOTTO_NUMBER_RANGE_ERROR;
-
 import java.util.Objects;
 
 public class LottoNumber implements Comparable<LottoNumber> {
 
+    public static final String LOTTO_NUMBER_RANGE_ERROR = "[ERROR] 로또 번호는 1부터 45까지 숫자입니다.";
     public static final int LOTTO_NUMBER_START = 1;
     public static final int LOTTO_NUMBER_END = 45;
     private final int value;
 
     public LottoNumber(int value) {
         if (value < LOTTO_NUMBER_START || value > LOTTO_NUMBER_END) {
-            throw new IllegalArgumentException(LOTTO_NUMBER_RANGE_ERROR.getMessage());
+            throw new IllegalArgumentException(LOTTO_NUMBER_RANGE_ERROR);
         }
         this.value = value;
     }
